# web/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BeautyMS API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/order/list/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Получение списка продуктов для заказа",
                "operationId": "e71370563a74f8502bcdba39275bfcae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список продуктов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "html": {
                                                        "type": "string"
                                                    },
                                                    "text": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order/time": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Получение доступного времени для заказа",
                "operationId": "93e5d23126cc806257dbb9aae505beaf",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Доступное время",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "text": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order/coupon/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Получение информации о купоне для продукта",
                "operationId": "f5f055021b0ae06da733fdf052d453fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML с информацией о купоне"
                    }
                }
            }
        },
        "/api/order/create": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Создание заказа",
                "operationId": "5dbc1c066366937e6fe5c039d10f469c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_coupon_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "street": {
                                        "type": "string"
                                    },
                                    "house": {
                                        "type": "string"
                                    },
                                    "apartment": {
                                        "type": "string"
                                    },
                                    "entrance": {
                                        "type": "string"
                                    },
                                    "floor": {
                                        "type": "string"
                                    },
                                    "intercom": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "time": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Созданный заказ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "phone": {
                                            "type": "string"
                                        },
                                        "city": {
                                            "type": "string"
                                        },
                                        "street": {
                                            "type": "string"
                                        },
                                        "house": {
                                            "type": "string"
                                        },
                                        "apartment": {
                                            "type": "string"
                                        },
                                        "entrance": {
                                            "type": "string"
                                        },
                                        "floor": {
                                            "type": "string"
                                        },
                                        "intercom": {
                                            "type": "string"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "time": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer"
                                        },
                                        "url": {
                                            "type": "string",
                                            "format": "uri"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса"
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order/item": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Получение информации о заказе",
                "operationId": "aacb1020a4e01f4bde3995074fe3d440",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Информация о заказе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "phone": {
                                            "type": "string"
                                        },
                                        "city": {
                                            "type": "string"
                                        },
                                        "street": {
                                            "type": "string"
                                        },
                                        "house": {
                                            "type": "string"
                                        },
                                        "apartment": {
                                            "type": "string"
                                        },
                                        "entrance": {
                                            "type": "string"
                                        },
                                        "floor": {
                                            "type": "string"
                                        },
                                        "intercom": {
                                            "type": "string"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "time": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса"
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order/order-list": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Получение списка заказов пользователя",
                "operationId": "7ad32b773c8863acd779f75bfcf70b88",
                "responses": {
                    "200": {
                        "description": "Список заказов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "city": {
                                                "type": "string"
                                            },
                                            "street": {
                                                "type": "string"
                                            },
                                            "house": {
                                                "type": "string"
                                            },
                                            "apartment": {
                                                "type": "string"
                                            },
                                            "entrance": {
                                                "type": "string"
                                            },
                                            "floor": {
                                                "type": "string"
                                            },
                                            "intercom": {
                                                "type": "string"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "time": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}